documentos en areas

+++ app/Http/Controllers/AreaController.php
@@ -23,9 +23,13 @@ class AreaController extends Controller
     public function store(Request $request)
     {
         ini_set('memory_limit','256M');
-        $extension = $request->file('document')->getClientOriginalExtension();
-        $path = $request->file('document')->storeAs('public/documents',$request->code.'.'.$extension);
-        $request->request->add(['path' => $path]);
+        if($request->document)
+        {
+            $extension = $request->file('document')->getClientOriginalExtension();
+            $path = $request->file('document')->storeAs('public/documents',$request->code.'.'.$extension);
+            $request->request->add(['path' => $path]);
+        }
+
         $area = \App\Area::create($request->all());
         return redirect('/areas');
     }
@@ -44,14 +48,14 @@ class AreaController extends Controller
     public function update(Request $request, Area $area)
     {
         ini_set('memory_limit','256M');
-
-        if ($area->path){
-            Storage::delete($area->path);
-        }
-
-        $extension = $request->file('document')->getClientOriginalExtension();
-        $path = $request->file('document')->storeAs('public/documents',$request->code.'.'.$extension);
-        $request->request->add(['path' => $path]);
+            if($request->document){
+                if ($area->path){
+                    Storage::delete($area->path);
+                }
+            $extension = $request->file('document')->getClientOriginalExtension();
+            $path = $request->file('document')->storeAs('public/documents',$request->code.'.'.$extension);
+            $request->request->add(['path' => $path]);
+            }
         $area->update($request->all());
         return redirect('/areas');
     }
